Reimplement PluginLoader - Add Plugin interface
To add a plugin, it needs to implement Plugin interface or one of
its sub-interfaces.

To have access to a List of implementations of a Plugin sub-interface
that sub-interface must be annotated with @Listable, and one can do it
through PluginLoader.getInstance().getPlugins(sub-interface-class)

To limit the number of implementations of a given Plugin sub-interface to
one, the @Listable annotation must set the isSingleInstance parameter to
true. @Listable(isSingleInstance = true)